refactor: separate release contract verification responsibilities
